Add files via upload
I designed the entire site and was also able to add a button that explains the rules of the game and gives a few tips. I didn't change the game.js file, but I did change the app.js file, so that I could make each button work, and so on.
Adham, I find that the game isn't very optimal. I've played several games against each AI, and there's a small problem with the scoring: on the first shot, the point is never counted, which reduces the score. I think that to have an almost perfect game, this is the last detail we're missing, and I'll try to fix it tonight in the last few hours.